+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 v1.5.4
 ------
+* Fix race condition which can cause the JVM to crash when the database is closed.
 * Fix deadlock during compaction caused by releasing the wrong node latch.
 * Don't reset checkpoint duration stat when thresholds aren't met.